server messages implemented
Implemented a system where I can communicate with the users using messages. A message is set up in the DB, the client checks for new messages every 5 mins (currently). If it finds a new message, it will print it in the console as well as in a toast notification. Such message will be only shown once.